Add inspection/analysis dialogs to CLOE data view
