+++ frontend/src/pages/DataViewPage/CloeDataTable.tsx
@@ -1,0 +1,87 @@
+import { Chip, Table, Typography } from '@equinor/eds-core-react'
+import { tokens } from '@equinor/eds-tokens'
+import { useLanguageContext } from 'components/Contexts/LanguageContext'
+import { StyledTable, StyledTableCell } from 'components/Styles/StyledComponents'
+import { DescriptionDisplay, TagIdDisplay } from 'components/Displays/TaskDisplay'
+import { Task } from 'models/Task'
+import { formatDateTime } from 'utils/StringFormatting'
+
+export const CloeDataTable = ({
+    tasks,
+    selectedTagId,
+    onSelectTag,
+}: {
+    tasks: Task[]
+    selectedTagId: string | null
+    onSelectTag: (tagId: string | null) => void
+}) => {
+    const { TranslateText } = useLanguageContext()
+
+    return (
+        <StyledTable>
+            <Table.Head>
+                <Table.Row>
+                    <StyledTableCell>#</StyledTableCell>
+                    <StyledTableCell>{TranslateText('Tag-ID')}</StyledTableCell>
+                    <StyledTableCell>{TranslateText('Description')}</StyledTableCell>
+                    <StyledTableCell>{TranslateText('Latest Value')}</StyledTableCell>
+                    <StyledTableCell>{TranslateText('Timestamp')}</StyledTableCell>
+                </Table.Row>
+            </Table.Head>
+            <Table.Body>
+                {tasks &&
+                    tasks.map((task, index) => {
+                        const isSelected = !!task.tagId && task.tagId === selectedTagId
+                        const taskHasWarning = !!task.inspection.analysisResult?.warning
+                        const backgroundColor = isSelected
+                            ? tokens.colors.interactive.primary__selected_highlight.hex
+                            : taskHasWarning
+                              ? tokens.colors.interactive.danger__highlight.hex
+                              : undefined
+
+                        return (
+                            <Table.Row
+                                key={task.id}
+                                onClick={() => {
+                                    if (!task.tagId) return
+                                    onSelectTag(isSelected ? null : task.tagId)
+                                }}
+                                style={{
+                                    cursor: task.tagId ? 'pointer' : 'default',
+                                    backgroundColor,
+                                }}
+                            >
+                                <Table.Cell>
+                                    <Chip>
+                                        <Typography variant="body_short_bold">{index + 1}</Typography>
+                                    </Chip>
+                                </Table.Cell>
+                                <Table.Cell>
+                                    <TagIdDisplay task={task} />
+                                </Table.Cell>
+                                <Table.Cell>
+                                    <DescriptionDisplay task={task} />
+                                </Table.Cell>
+                                {task.inspection.analysisResult?.value ? (
+                                    <Table.Cell>
+                                        <Typography>
+                                            {Math.round(parseFloat(task.inspection.analysisResult?.value)) + '%'}
+                                        </Typography>
+                                    </Table.Cell>
+                                ) : (
+                                    <Table.Cell>
+                                        <Typography>{TranslateText('Not available')}</Typography>
+                                    </Table.Cell>
+                                )}
+                                {task.endTime && (
+                                    <Table.Cell>
+                                        <Typography>{formatDateTime(task.endTime, 'dd.MM.yy - HH:mm')}</Typography>
+                                    </Table.Cell>
+                                )}
+                            </Table.Row>
+                        )
+                    })}
+            </Table.Body>
+        </StyledTable>
+    )
+}
