+++ frontend/src/pages/DataViewPage/CloeDataViewPage.tsx
@@ -1,27 +1,24 @@
-import { Button, Chip, Table, Typography } from '@equinor/eds-core-react'
+import { Button, Typography } from '@equinor/eds-core-react'
 import { Mission, MissionStatus } from 'models/Mission'
 import { useContext, useEffect, useMemo, useState } from 'react'
 import styled from 'styled-components'
 import { tokens } from '@equinor/eds-tokens'
 import { useLanguageContext } from 'components/Contexts/LanguageContext'
-import {
-    StyledCardsWidth,
-    StyledPage,
-    StyledTable,
-    StyledTableAndMap,
-    StyledTableCell,
-} from 'components/Styles/StyledComponents'
+import { StyledPage, StyledTableAndMap } from 'components/Styles/StyledComponents'
 import { SignalREventLabels, useSignalRContext } from 'components/Contexts/SignalRContext'
 import { useBackendApi } from 'api/UseBackendApi'
 import { InstallationContext } from 'components/Contexts/InstallationContext'
 import { Task } from 'models/Task'
-import { formatDateTime } from 'utils/StringFormatting'
 import { TimeseriesLinePlot, TimeseriesLinePlotData } from 'components/Displays/TimeseriesLinePlot'
-import { DescriptionDisplay, TagIdDisplay } from 'components/Displays/TaskDisplay'
 import { PlantMap } from 'pages/MissionPage/MapPosition/PointillaMapView'
+import { AnalysisOverviewSection, InspectionOverviewSection } from 'pages/InspectionReportPage/ImageOverview'
 import { StyledImagesSection } from 'pages/InspectionReportPage/InspectionStyles'
 import { InspectionImageWithPlaceholder } from 'pages/InspectionReportPage/InspectionReportImage'
 import { AnalysisResultDialogContent } from 'pages/MissionPage/AnalysisResultView'
+import { InspectionDialogView } from 'pages/InspectionReportPage/InspectionView'
+import { AnalysisResultDialogView } from 'pages/MissionPage/AnalysisResultView'
+import { useSearchParams } from 'react-router-dom'
+import { CloeDataTable } from './CloeDataTable'
 
 const CloeMapWrapper = styled.div`
     .leaflet-tooltip.circleLabel {
@@ -30,8 +27,15 @@ const CloeMapWrapper = styled.div`
         border-radius: 2px !important;
     }
 `
-const CloeTableWrapper = styled.div`
+const WhiteBackgroundBand = styled.div`
+    display: flex;
+    flex-direction: column;
     align-self: flex-start;
+    margin: 0 -3rem;
+    padding: 24px 3rem;
+    gap: 24px;
+    width: 100%;
+    background: ${tokens.colors.ui.background__default.hex};
 `
 const TimeRangeToggle = styled.div`
     display: inline-flex;
@@ -39,20 +43,26 @@ const TimeRangeToggle = styled.div`
     gap: 4px;
     padding: 4px;
     border-radius: 6px;
-    background-color: ${tokens.colors.ui.background__light.hex};
     box-shadow: inset 0 0 0 1px ${tokens.colors.ui.background__medium.hex};
 `
 const TimeRangeToggleButton = styled(Button)`
     border-radius: 4px;
 `
+const CloeChartArea = styled.div`
+    display: flex;
+    flex-direction: column;
+    max-width: 1250px;
+    gap: 15px;
+`
 const StyledTopAlignedImagesSection = styled(StyledImagesSection)`
     align-items: flex-start;
+    gap: 40px;
 `
 const StyledCloeImageCard = styled.div`
     display: flex;
     flex-direction: column;
     gap: 8px;
-    max-width: 480px;
+    max-width: 530px;
 `
 
 enum TimeRange {
@@ -77,86 +87,6 @@ const checkIfAnalysableDescription = (description: string) => {
     return Object.values(CloeAnalysableDescriptions).includes(description as CloeAnalysableDescriptions)
 }
 
-const CloeDataTable = ({
-    tasks,
-    selectedTagId,
-    onSelectTag,
-}: {
-    tasks: Task[]
-    selectedTagId: string | null
-    onSelectTag: (tagId: string | null) => void
-}) => {
-    const { TranslateText } = useLanguageContext()
-
-    return (
-        <StyledTable>
-            <Table.Head>
-                <Table.Row>
-                    <StyledTableCell>#</StyledTableCell>
-                    <StyledTableCell>{TranslateText('Tag-ID')}</StyledTableCell>
-                    <StyledTableCell>{TranslateText('Description')}</StyledTableCell>
-                    <StyledTableCell>{TranslateText('Latest Value')}</StyledTableCell>
-                    <StyledTableCell>{TranslateText('Timestamp')}</StyledTableCell>
-                </Table.Row>
-            </Table.Head>
-            <Table.Body>
-                {tasks &&
-                    tasks.map((task, index) => {
-                        const isSelected = !!task.tagId && task.tagId === selectedTagId
-                        const taskHasWarning = !!task.inspection.analysisResult?.warning
-                        const backgroundColor = isSelected
-                            ? tokens.colors.interactive.primary__selected_highlight.rgba
-                            : taskHasWarning
-                              ? tokens.colors.interactive.danger__highlight.rgba
-                              : undefined
-
-                        return (
-                            <Table.Row
-                                key={task.id}
-                                onClick={() => {
-                                    if (!task.tagId) return
-                                    onSelectTag(isSelected ? null : task.tagId)
-                                }}
-                                style={{
-                                    cursor: task.tagId ? 'pointer' : 'default',
-                                    backgroundColor,
-                                }}
-                            >
-                                <Table.Cell>
-                                    <Chip>
-                                        <Typography variant="body_short_bold">{index + 1}</Typography>
-                                    </Chip>
-                                </Table.Cell>
-                                <Table.Cell>
-                                    <TagIdDisplay task={task} />
-                                </Table.Cell>
-                                <Table.Cell>
-                                    <DescriptionDisplay task={task} />
-                                </Table.Cell>
-                                {task.inspection.analysisResult?.value ? (
-                                    <Table.Cell>
-                                        <Typography>
-                                            {Math.round(parseFloat(task.inspection.analysisResult?.value)) + '%'}
-                                        </Typography>
-                                    </Table.Cell>
-                                ) : (
-                                    <Table.Cell>
-                                        <Typography>{TranslateText('Analysis result not available')}</Typography>
-                                    </Table.Cell>
-                                )}
-                                {task.endTime && (
-                                    <Table.Cell>
-                                        <Typography>{formatDateTime(task.endTime, 'dd.MM.yy - HH:mm')}</Typography>
-                                    </Table.Cell>
-                                )}
-                            </Table.Row>
-                        )
-                    })}
-            </Table.Body>
-        </StyledTable>
-    )
-}
-
 export const CloeDataViewPage = () => {
     const { TranslateText } = useLanguageContext()
     const { installation } = useContext(InstallationContext)
@@ -166,6 +96,9 @@ export const CloeDataViewPage = () => {
     const [selectedTagId, setSelectedTagId] = useState<string | null>(null)
     const [selectedDataPoint, setSelectedDataPoint] = useState<{ tagId: string; timeMs: number } | null>(null)
     const backendApi = useBackendApi()
+    const [searchParams] = useSearchParams()
+    const inspectionId = searchParams.get('inspectionId') ?? undefined
+    const analysisId = searchParams.get('analysisId') ?? undefined
 
     const fetchMissions = (): Promise<Mission[]> => {
         const minEndTime = Math.floor((Date.now() - CLOE_FETCH_WINDOW_DAYS * MS_PER_DAY) / MS_PER_SECOND)
@@ -283,11 +216,11 @@ export const CloeDataViewPage = () => {
 
     return (
         <StyledPage>
-            <StyledCardsWidth>
-                <Typography variant="h2">{TranslateText('Data View for Constant Level Oilers')}</Typography>
-                <StyledTableAndMap>
-                    {latestCloeMission && (
-                        <CloeTableWrapper>
+            <Typography variant="h2">{TranslateText('Data View for Constant Level Oilers')}</Typography>
+            {(latestCloeMission || (plantCode && mapMission)) && (
+                <WhiteBackgroundBand>
+                    <StyledTableAndMap>
+                        {latestCloeMission && (
                             <CloeDataTable
                                 tasks={latestCloeMission.tasks}
                                 selectedTagId={selectedTagId}
@@ -296,88 +229,100 @@ export const CloeDataViewPage = () => {
                                     setSelectedDataPoint(null)
                                 }}
                             />
-                        </CloeTableWrapper>
-                    )}
-                    {plantCode && mapMission && (
-                        <CloeMapWrapper>
-                            <PlantMap
-                                key={selectedTagId ?? 'all'}
-                                plantCode={plantCode}
-                                floorId="0"
-                                mission={mapMission}
-                            />
-                        </CloeMapWrapper>
-                    )}
-                </StyledTableAndMap>
-                {selectedTask && selectedTask.inspection.isarInspectionId && (
-                    <StyledTopAlignedImagesSection>
+                        )}
+                        {plantCode && mapMission && (
+                            <CloeMapWrapper>
+                                <PlantMap
+                                    key={selectedTagId ?? 'all'}
+                                    plantCode={plantCode}
+                                    floorId="0"
+                                    mission={mapMission}
+                                />
+                            </CloeMapWrapper>
+                        )}
+                    </StyledTableAndMap>
+                </WhiteBackgroundBand>
+            )}
+            {selectedTask && selectedTask.inspection.isarInspectionId && (
+                <StyledTopAlignedImagesSection>
+                    <StyledCloeImageCard>
+                        <Typography variant="h4">{inspectionImageTitle}</Typography>
+                        <InspectionImageWithPlaceholder task={selectedTask} isLargeImage={true} />
+                    </StyledCloeImageCard>
+                    {selectedTask.inspection.analysisResult?.storageAccount && (
                         <StyledCloeImageCard>
-                            <Typography variant="h4">{inspectionImageTitle}</Typography>
-                            <InspectionImageWithPlaceholder task={selectedTask} isLargeImage={true} />
+                            <Typography variant="h4">{analysisImageTitle}</Typography>
+                            <AnalysisResultDialogContent currentTask={selectedTask} />
                         </StyledCloeImageCard>
-                        {selectedTask.inspection.analysisResult?.storageAccount && (
-                            <StyledCloeImageCard>
-                                <Typography variant="h4">{analysisImageTitle}</Typography>
-                                <AnalysisResultDialogContent currentTask={selectedTask} />
-                            </StyledCloeImageCard>
-                        )}
-                    </StyledTopAlignedImagesSection>
-                )}
-                {cloeMissions.length > 0 && (
-                    <>
-                        <Typography variant="h3">{TranslateText('Estimated oil level')}</Typography>
-                        <TimeRangeToggle role="group" aria-label={TranslateText('Measured oil level')}>
-                            <TimeRangeToggleButton
-                                variant={timeRange === TimeRange.SevenDays ? 'contained' : 'ghost'}
-                                aria-pressed={timeRange === TimeRange.SevenDays}
-                                onClick={() => {
-                                    setTimeRange(TimeRange.SevenDays)
-                                    setSelectedDataPoint(null)
-                                }}
-                            >
-                                {TranslateText('7 days')}
-                            </TimeRangeToggleButton>
-                            <TimeRangeToggleButton
-                                variant={timeRange === TimeRange.OneMonth ? 'contained' : 'ghost'}
-                                aria-pressed={timeRange === TimeRange.OneMonth}
-                                onClick={() => {
-                                    setTimeRange(TimeRange.OneMonth)
-                                    setSelectedDataPoint(null)
-                                }}
-                            >
-                                {TranslateText('1 month')}
-                            </TimeRangeToggleButton>
-                        </TimeRangeToggle>
-                        {recentSphericalGlassMissions.length > 0 ? (
-                            <TimeseriesLinePlot
-                                data={linePlotData}
-                                yLabel={TranslateText('Fill [%]')}
-                                ymin={0}
-                                ymax={100}
-                                selectedPoint={
-                                    selectedDataPoint
-                                        ? {
-                                              id: selectedDataPoint.tagId,
-                                              time: new Date(selectedDataPoint.timeMs),
-                                          }
-                                        : undefined
-                                }
-                                onPointClick={({ id, time }) => {
-                                    const timeMs = time.getTime()
-                                    setSelectedDataPoint((current) =>
-                                        current && current.tagId === id && current.timeMs === timeMs
-                                            ? null
-                                            : { tagId: id, timeMs }
-                                    )
-                                    setSelectedTagId(id)
-                                }}
-                            />
-                        ) : (
-                            <Typography>{TranslateText('No data available in the selected time range')}</Typography>
-                        )}
-                    </>
-                )}
-            </StyledCardsWidth>
+                    )}
+                </StyledTopAlignedImagesSection>
+            )}
+            {cloeMissions.length > 0 && (
+                <CloeChartArea>
+                    <Typography variant="h3">{TranslateText('Estimated oil level')}</Typography>
+                    <TimeRangeToggle role="group" aria-label={TranslateText('Measured oil level')}>
+                        <TimeRangeToggleButton
+                            variant={timeRange === TimeRange.SevenDays ? 'contained' : 'ghost'}
+                            aria-pressed={timeRange === TimeRange.SevenDays}
+                            onClick={() => {
+                                setTimeRange(TimeRange.SevenDays)
+                                setSelectedDataPoint(null)
+                            }}
+                        >
+                            {TranslateText('7 days')}
+                        </TimeRangeToggleButton>
+                        <TimeRangeToggleButton
+                            variant={timeRange === TimeRange.OneMonth ? 'contained' : 'ghost'}
+                            aria-pressed={timeRange === TimeRange.OneMonth}
+                            onClick={() => {
+                                setTimeRange(TimeRange.OneMonth)
+                                setSelectedDataPoint(null)
+                            }}
+                        >
+                            {TranslateText('1 month')}
+                        </TimeRangeToggleButton>
+                    </TimeRangeToggle>
+                    {recentSphericalGlassMissions.length > 0 ? (
+                        <TimeseriesLinePlot
+                            data={linePlotData}
+                            yLabel={TranslateText('Fill [%]')}
+                            ymin={0}
+                            ymax={100}
+                            selectedPoint={
+                                selectedDataPoint
+                                    ? {
+                                          id: selectedDataPoint.tagId,
+                                          time: new Date(selectedDataPoint.timeMs),
+                                      }
+                                    : undefined
+                            }
+                            onPointClick={({ id, time }) => {
+                                const timeMs = time.getTime()
+                                setSelectedDataPoint((current) =>
+                                    current && current.tagId === id && current.timeMs === timeMs
+                                        ? null
+                                        : { tagId: id, timeMs }
+                                )
+                                setSelectedTagId(id)
+                            }}
+                        />
+                    ) : (
+                        <Typography>{TranslateText('No data available in the selected time range')}</Typography>
+                    )}
+                </CloeChartArea>
+            )}
+            {latestCloeMission && !selectedDataPoint && (
+                <WhiteBackgroundBand>
+                    <InspectionOverviewSection tasks={latestCloeMission.tasks} />
+                    <AnalysisOverviewSection tasks={latestCloeMission.tasks} />
+                </WhiteBackgroundBand>
+            )}
+            {latestCloeMission && inspectionId && !selectedDataPoint && (
+                <InspectionDialogView selectedInspectionId={inspectionId} tasks={latestCloeMission.tasks} />
+            )}
+            {latestCloeMission && analysisId && !selectedDataPoint && (
+                <AnalysisResultDialogView selectedAnalysisId={analysisId} tasks={latestCloeMission.tasks} />
+            )}
         </StyledPage>
     )
 }
