+++ frontend/src/pages/MissionPage/MissionPage.tsx
@@ -12,7 +12,7 @@ import { AlertCategory } from 'components/Alerts/AlertsBanner'
 import { useMediaStreamContext } from 'components/Contexts/MediaStreamContext'
 import { StyledCardsWidth, VideoStreamSection } from 'components/Styles/StyledComponents'
 import { InspectionDialogView } from '../InspectionReportPage/InspectionView'
-import { InspectionOverviewSection } from '../InspectionReportPage/InspectionOverview'
+import { InspectionOverviewSection } from '../InspectionReportPage/ImageOverview'
 import { TaskTableAndMap } from './TaskTableAndMap'
 import { AnalysisResultDialogView } from './AnalysisResultView'
 import { tokens } from '@equinor/eds-tokens'
